trying to get width of ColorBar component in pixels
